Expose dialog events to parent component
It's not a good idea to track the visibility of the dialog in the state,
because it will go out of sync with the state attribute used to show or
hide it in the parent component (i.e. editedActivity or error).

Now that the parent component is notified that the dialogs have been
dismissed, it can update its own state, which in turn hides the dialogs.

One minor downside is that the dialogs are recreated every time they are
shown. For the edit dialog it actually simplifies the focus logic.